Add genie:proc_name/1,2 for formatting process names
genie:proc_name/1 will return the name the process is registered as
or the pid of the process, e.g. {local, LocalName} returns LocalName.

genie:proc_name/2 is as above except the second argument takes an
options list. If this list contains the atom verify then it is check if
the calling process is registered with the name given. If not the
function exits.