Add genie:cast/3 and genie:send/3
genie:cast/3 acts the same as gen_server:cast/3 except that if a global
or custom registered name has no associated pid it does not raise an
error. This makes it consistent with its handling of when a locally
registered name has no associated pid.

genie:send/3 acts the same as gen_fsm:send_event/2 except that if
a global or custom registered name as no associated pid it raises an
error. This makes it consistent with its handling of when a locally
registered name has no associated pid.

This means genie:cast/3 messages may arrive out of order when sending to
another node, will return immediately and should never raise an error.
genie:send/3 messages will arrive in order when sending to another node,
may not return immediately (as will block to connect to another node)
and will raise an error when a registered name has no associated pid -
except when sending to a locally registered pid on another node.