Add genie:call_list/3,4.
genie:call_list/3,4 makes parallel calls to a list of processes. It
behaves in the same way to genie:call/3,4 except that a tuple is
returned with a list of replies and a list of errors. The pid in the
reference of the call message will be a middleman's pid.